+++ modules/article/views/items/_form.php
@@ -1,11 +1,11 @@
 <?php
 use yii\easyii\helpers\Image;
+use yii\easyii\models\Setting;
 use yii\easyii\widgets\DateTimePicker;
 use yii\easyii\widgets\TagsInput;
 use yii\helpers\Html;
 use yii\helpers\Url;
 use yii\widgets\ActiveForm;
-use yii\easyii\widgets\Redactor;
 use yii\easyii\widgets\SeoForm;
 
 $module = $this->context->module->id;
@@ -28,12 +28,15 @@ $module = $this->context->module->id;
     <?= $form->field($model, 'short')->textarea() ?>
 <?php endif; ?>
 
-<?= $form->field($model, 'text')->widget(Redactor::className(),[
-    'options' => [
+<?= $form->field($model, 'text')->widget(\vova07\imperavi\Widget::className(), [
+    'settings' => [
+        'lang' => \yii\easyii\helpers\Data::getLocale(),
         'minHeight' => 400,
-        'imageUpload' => Url::to(['/admin/redactor/upload', 'dir' => 'article'], true),
-        'fileUpload' => Url::to(['/admin/redactor/upload', 'dir' => 'article'], true),
-        'plugins' => ['fullscreen']
+        'imageUpload' => Url::to(['/admin/redactor/image-upload']),
+        'fileUpload' => Url::to(['/admin/redactor/file-upload']),
+        'imageManagerJson' => Url::to(['/admin/redactor/images-get']),
+        'fileManagerJson' => Url::to(['/admin/redactor/files-get']),
+        'plugins' => Setting::getAsArray('redactor_plugins')
     ]
 ]) ?>