+++ modules/faq/views/a/_form.php
@@ -1,24 +1,34 @@
 <?php
+use yii\easyii\models\Setting;
 use yii\easyii\widgets\TagsInput;
 use yii\helpers\Html;
+use yii\helpers\Url;
 use yii\widgets\ActiveForm;
-use yii\easyii\widgets\Redactor;
+
 ?>
 <?php $form = ActiveForm::begin([
     'options' => ['class' => 'model-form']
 ]); ?>
-<?= $form->field($model, 'question')->widget(Redactor::className(),[
-    'options' => [
-        'minHeight' => 300,
-        'buttons' => ['bold', 'italic', 'unorderedlist', 'link'],
-        'linebreaks' => true
+<?= $form->field($model, 'question')->widget(\vova07\imperavi\Widget::className(), [
+    'settings' => [
+        'lang' => \yii\easyii\helpers\Data::getLocale(),
+        'minHeight' => 400,
+        'imageUpload' => Url::to(['/admin/redactor/image-upload']),
+        'fileUpload' => Url::to(['/admin/redactor/file-upload']),
+        'imageManagerJson' => Url::to(['/admin/redactor/images-get']),
+        'fileManagerJson' => Url::to(['/admin/redactor/files-get']),
+        'plugins' => Setting::getAsArray('redactor_plugins')
     ]
 ]) ?>
-<?= $form->field($model, 'answer')->widget(Redactor::className(),[
-    'options' => [
-        'minHeight' => 300,
-        'buttons' => ['bold', 'italic', 'unorderedlist', 'link'],
-        'linebreaks' => true
+<?= $form->field($model, 'answer')->widget(\vova07\imperavi\Widget::className(), [
+    'settings' => [
+        'lang' => \yii\easyii\helpers\Data::getLocale(),
+        'minHeight' => 400,
+        'imageUpload' => Url::to(['/admin/redactor/image-upload']),
+        'fileUpload' => Url::to(['/admin/redactor/file-upload']),
+        'imageManagerJson' => Url::to(['/admin/redactor/images-get']),
+        'fileManagerJson' => Url::to(['/admin/redactor/files-get']),
+        'plugins' => Setting::getAsArray('redactor_plugins')
     ]
 ]) ?>