+++ modules/page/views/a/_form.php
@@ -1,8 +1,8 @@
 <?php
+use yii\easyii\models\Setting;
 use yii\helpers\Html;
 use yii\helpers\Url;
 use yii\widgets\ActiveForm;
-use yii\easyii\widgets\Redactor;
 use yii\easyii\widgets\SeoForm;
 ?>
 <?php $form = ActiveForm::begin([
@@ -10,12 +10,15 @@ use yii\easyii\widgets\SeoForm;
     'options' => ['class' => 'model-form']
 ]); ?>
 <?= $form->field($model, 'title') ?>
-<?= $form->field($model, 'text')->widget(Redactor::className(),[
-    'options' => [
+<?= $form->field($model, 'text')->widget(\vova07\imperavi\Widget::className(), [
+    'settings' => [
+        'lang' => \yii\easyii\helpers\Data::getLocale(),
         'minHeight' => 400,
-        'imageUpload' => Url::to(['/admin/redactor/upload', 'dir' => 'pages']),
-        'fileUpload' => Url::to(['/admin/redactor/upload', 'dir' => 'pages']),
-        'plugins' => ['fullscreen']
+        'imageUpload' => Url::to(['/admin/redactor/image-upload']),
+        'fileUpload' => Url::to(['/admin/redactor/file-upload']),
+        'imageManagerJson' => Url::to(['/admin/redactor/images-get']),
+        'fileManagerJson' => Url::to(['/admin/redactor/files-get']),
+        'plugins' => Setting::getAsArray('redactor_plugins')
     ]
 ]) ?>